pyproject.toml
[project]
name = "evoman"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "deap>=1.4.1",
    "matplotlib>=3.9.2",
    "neat-python>=0.92",
    "numpy>=2.1.1",
    "pygame>=2.6.0",
    "pyparsing==3.1.3",
]

[tool.pyright]
include = ["."]
exclude = [ ".venv" ]
venvPath =  "."
venv = ".venv"
